refactor(utils/hooks): Refactor utility hooks
The 'useGetLaundryServices.tsx' file in the 'utils/hooks' directory was modified to improve code clarity and reusability by accepting a 'type' parameter and conditionally fetching laundry services based on the type ('admin' or 'web'). Additionally, the refetch interval for 'useGetLaundryServices' was updated to 10 seconds for 'web' type and 5 seconds for 'admin' type. In the 'useNotification.tsx' file, the default duration for notifications was reduced to 3333 milliseconds for a better user experience.